refactor(daemon): extract transcript capture helpers
